# example/runtime-guard/express/openapi.flow.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Runtime Guard Express Demo API",
    "version": "1.0.0"
  },
  "paths": {
    "/payments": {
      "post": {
        "operationId": "createPayment",
        "summary": "Create payment",
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment created"
          }
        },
        "x-openapi-flow": {
          "version": "1.0",
          "id": "create-payment-flow",
          "current_state": "AUTHORIZED",
          "transitions": [
            {
              "target_state": "CAPTURED",
              "trigger_type": "synchronous",
              "next_operation_id": "capturePayment"
            }
          ]
        }
      }
    },
    "/payments/{id}/capture": {
      "post": {
        "operationId": "capturePayment",
        "summary": "Capture payment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment captured"
          }
        },
        "x-openapi-flow": {
          "version": "1.0",
          "id": "capture-payment-flow",
          "current_state": "CAPTURED",
          "transitions": []
        }
      }
    },
    "/payments/{id}": {
      "get": {
        "operationId": "getPayment",
        "summary": "Get payment",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Payment state"
          },
          "404": {
            "description": "Payment not found"
          }
        }
      }
    }
  }
}